requirements.txt
fastapi>=0.104.0,<1.0
uvicorn>=0.24.0,<1.0
pydantic>=2.5.0,<3.0
python-dotenv>=1.0.0,<2.0
pynacl>=1.5.0,<2.0
requests>=2.31.0,<3.0
pytest>=7.4.0,<9.0
